Validate Plot Inputs

This internal function validates the inputs for plotting functions to ensure proper data format and parameter values.

Usage

.validatePlotYXsXInputs(
  data,
  metaData,
  geomPointAttributes,
  geomErrorbarAttributes,
  geomGuestLineAttributes,
  geomComparisonLineAttributes,
  geomLLOQAttributes,
  groupAesthetics,
  comparisonLineVector,
  addRegression,
  addGuestLimits,
  deltaGuest,
  residualScale,
  asSquarePlot,
  xscale,
  xscale.args,
  yscale,
  yscale.args,
  observedDataDirection
)

Arguments

data

A data.frame containing the data to plot.

metaData

A named list of information about data such as the dimension and unit of its variables.

geomPointAttributes

A list with arguments which are passed on to the call ggplot2::geom_point

geomErrorbarAttributes

A list with arguments which are passed on to the call ggplot2::geom_errorbar

geomGuestLineAttributes

A list of arguments passed to ggplot2::geom_function to display guest criteria.

geomComparisonLineAttributes

A list of arguments passed to ggplot2::hline or ggplot2::abline to display comparison lines.

geomLLOQAttributes

A list with arguments which are passed on to the call ggplot2::geom_hline

groupAesthetics

vector of aesthetics, which are used for columns mapped with groupby,

comparisonLineVector

A vector defining the comparison lines.

addRegression

A boolean that activates the insertion of a regression line.

addGuestLimits

A boolean that activates the insertion of guest limits.

deltaGuest

Numeric value parameter for the Guest function.

residualScale

Either "linear", "log", or "ratio" scale for residuals.

asSquarePlot

A boolean; if true, the plot is returned as a square plot with aspect ratio = 1 and fixed ratios.

xscale

either 'linear' then ggplot2::scale_x_continuous() or 'log' then ggplot2::scale_x_log10() is used

xscale.args

list of arguments passed to ggplot2::scale_x_continuous() or ggplot2::scale_x_log10()

yscale

either 'linear' then ggplot2::scale_y_continuous() or 'log' then ggplot2::scale_y_log10() is used

yscale.args

list of arguments passed to ggplot2::scale_y_continuous() or ggplot2::scale_y_log10()

observedDataDirection

Either 'x' or 'y', defining the direction of observed data.

Value

Invisible NULL if validation is successful; otherwise, an error is raised.