Create plot function for ggWatermark.

This method allows for the ggWatermark object to be plotted using the standard print method.

Usage

# S3 method for class 'ggWatermark'
plot(x, ...)

Arguments

x

A ggWatermark object to be printed.

...

Additional arguments to be passed to the print method.

See also