generates residual quantile quantile plot

For details and examples see the vignettes:

vignette("Goodness of fit", package = "ospsuite.plots")

vignette("ospsuite.plots", package = "ospsuite.plots")

Usage

plotQQ(
  data,
  mapping,
  metaData = NULL,
  xscale.args = list(),
  residualScale = ResidualScales$log,
  yscale.args = list(),
  geomQQAttributes = list(),
  geomQQLineAttributes = geomQQAttributes,
  groupAesthetics = c("colour", "fill", "shape")
)

Arguments

data

´data.frame` with data to plot

mapping

a list of aesthetic mappings to use for plot, additional to {ggplot2} aesthetics, the aesthetics groupby,error,error_relative,lloq, mdv, y2axis are available, see vignettes for more details and examples

metaData

A named list of information about data such as the dimension and unit of its variables.

xscale.args

list of arguments passed to ggplot2::scale_x_continuous() or ggplot2::scale_x_log10()

residualScale

Either "linear" or "log" for scaling residuals. For linear: residuals = observed - predicted. For log: residuals = log(observed) - log(predicted). The y-axis scale remains linear in both cases.

yscale.args

list of arguments passed to ggplot2::scale_y_continuous() or ggplot2::scale_y_log10()

geomQQAttributes

A list of arguments passed to ggplot2::stat_qq().

geomQQLineAttributes

A list of arguments passed to ggplot2::stat_qq_line().

groupAesthetics

A character vector of aesthetic names used for grouping data points in the Q-Q plot. Common options include "colour", "fill", "shape", "linetype", and "size".

Value

A ggplot object

See also