Geom to point unicode as shapes

Define a Geom using ggplot2::ggproto() and based on GeomPoint. The Geom internally uses textGrob instead of pointsGrob so that fonts leverage for drawing shapes. The grid and scales packages are supposed to be required by ggplot2. So there should not be any issue as installing ggplot2 should install those 2 packages.

Usage

GeomPointUnicodeProto

Format

An object of class GeomPointUnicodeProto (inherits from GeomPoint, Geom, ggproto, gg) of length 4.